FEAT: separated the credit transaction logic into a redis lua scripts which the credit service calls using the credits-utils. so we changed the redis entrypoint to log in using the first admin credentials (so it works in prod mode when default user is off)
FIX: added a flag that prevent unecessary call of reset-hard-cost of the functions

REFAC:  changed the definition of the cost. The notifyr app sends a bills after each request. hence the changement to bill. The bills are stored inside a redis list the user already paid, but each week the bills are squash into a receipts which is the summary of the purchase made in a week

CHORES: updated the permission to call the lua scripts.