FEAT: added a huge segmentation between in the deployment, now each host redis,rabbit,postgres is either internally or externally, the vault needs to be provided a segmentation account that will hold data specif the notifyr, so the dynamic secret only affects this data. if a specific host is not set it will be deployed internally. whether the database is set in/ex vault will issue dynamic credentials. So password must always be set at the first deployment
FEAT: the notifyr environnement must have its own internal vault (manage dynamic secrets) and redis instance (manage credits and events) to run core code. But they can be another external redis instance which will result either way in another root redis vault config, which can point to either the same internal host or one int and another ext.

REFAC: from those changes, we needed segmented security in case the redis host is external which is wil be the same as if it internal, but this idea raised the change of the a certain keyprefix that is authorization command that can be applied to redis. and only the Redis associated will celery will suffer from this changes, so i added a global prefix on anything related to redis_celery db ( in/ext)

CHORES: removed the base .env in the docker services, so the .env will be the only .env file needed for notifyr since all the configuration is too heavy at the moment only relevant env is needed in the docker compose and can actually changes ?Notifyr behavior